% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.bayesTestResults}
\alias{plot.bayesTestResults}
\title{plot method for bayesTestResults object}
\usage{
\method{plot}{bayesTestResults}(
  bayesTestTable,
  xAxis = .data$topScorerSampleSize,
  yAxis = .data$lowScorerSampleSize,
  fill = .data$probability,
  ...
)
}
\arguments{
\item{bayesTestTable}{output from \code{\link{runABbayesTest}}}

\item{xAxis}{unquoted column name containing the data to plot on the x axis of the tile plot}

\item{yAxis}{unquoted column name containing the data to plot on the y axis of the tile plot}

\item{fill}{unquoted column name containing the data to file the tile plot}
}
\description{
plot method for bayesTestResults object
}
