% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.powerTestResults}
\alias{plot.powerTestResults}
\title{plot method for powerTestResults object}
\usage{
\method{plot}{powerTestResults}(
  powerTestTable,
  xAxis = .data$topScorerSampleSize,
  yAxis = .data$lowScorerSampleSize,
  fill = .data$power,
  ...
)
}
\arguments{
\item{bayesTestTable}{output from \code{\link{runFrequentistPower}}}
}
\description{
plot method for powerTestResults object
}
