% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomSample.R
\name{randomSample}
\alias{randomSample}
\alias{randomSample.bernoulli}
\title{Random generation for some data distribution}
\usage{
randomSample(...)

\method{randomSample}{bernoulli}(n, size, prob, ...)
}
\arguments{
\item{...}{S3 method compatibility}

\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{size}{number of trials (zero or more).}

\item{prob}{probability of success on each trial.}
}
\description{
Random generation for some data distribution
}
