% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianMethods.R
\name{runABbayesTest}
\alias{runABbayesTest}
\title{bayesTest wrapper}
\usage{
runABbayesTest(
  topScorerHitRate,
  topScorerSampleSize,
  lowScorerhitRate,
  lowScorerSampleSize,
  priors,
  distribution = c("bernoulli", "bernoulliC"),
  nSimulatedData = 10,
  n_samples = 1e+05
)
}
\arguments{
\item{topScorerHitRate}{Numeric, hit rate for the top scorer}

\item{topScorerSampleSize}{Named numeric vector, group size for the top scorer}

\item{lowScorerhitRate}{Numeric, hit rate for the low scorer}

\item{lowScorerSampleSize}{Named numeric vector, group size for the low scorer}

\item{priors}{Named vector or named list providing priors as required by the specified distribution:
\itemize{
          \item     For 'bernoulli' distribution \code{list("alpha" = val1, "beta" = val2)}
          \item     For 'normal' distribution \code{c("mu" = val1, "lambda" = val2, "alpha" = val3, "beta" = val4)}
          \item     For 'lognormal' distribution \code{c("mu" = val1, "lambda" = val2, "alpha" = val3, "beta" = val4)}
          \item     For 'poisson' distribution \code{c("shape" = val1, "rate" = val2)}
          \item     For 'exponential' distribution \code{list("shape" = val1, "rate" = val2)}
          \item     For 'uniform' distribution \code{c("xm" = val1, "alpha" = val2)}
          \item     For 'bernoulliC' distribution: same prior definitions as 'bernoulli'
          \item     For 'poissonC' distribution: same prior definitions as 'poisson'
          }

          See \link[bayesAB]{plotDistributions} or the \emph{Note} section of this help document for more info.}

\item{distribution}{of underlying A/B test data. See more in details}

\item{nSimulatedData}{number of trials for the simulated data}

\item{n_samples}{Number of posterior samples to draw. Should be large enough for the distribution to converge. 1e5 is a good rule of thumb.
Not used for closed form tests.}
}
\value{
Returns bayesTest results in a tidy table

\describe{
\item{topScorerSampleSize}{numeric - top-scorer sample size}
\item{lowScorerSampleSize}{numeric - low-scorer sample size}
\item{probability}{numeric - Probability of top-scorer hit rate to be better than low-scorer hit rate P(topScorerHitRate > lowScorerhitRate)}
\item{posteriorTopScorerData}{list - distribution over the posterior for top-scorer (if available)}
\item{posteriorLowScorerData}{list - distribution over the posterior for low-scorer (if available)}
}
}
\description{
Wrapper to run multiple bayesian A/B tests on lists of A_data, B_data and priors.
}
\details{
\itemize{
\item Only \code{bernoulli} & \code{bernoulliC} distribution are supported for now but the user can easily add a \code{tidyBayesTest} S3 method for the distribution
that suits better their data. The distribution that can be implemented are the one available in the bayesAB package see \code{\link[bayesAB]{bayesTest}}.
\item Bernoulli can be used if the data is binary, modeled by 1s and 0s, according to a specific probability p of a 1 occurring. The conjugate Beta distribution for the parameter p should to be used.
\item \code{bernoulli} is using Monte Carlo sampling to estimate the posterior probability while \code{BernoulliC} is computing the posterior probability analytically (faster but doesn't report the distribution over the posterior)
}
}
