% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateData}
\alias{simulateData}
\alias{simulateData.bernoulli}
\alias{simulateData.bernoulliC}
\title{Random generation for some data distribution}
\usage{
simulateData(...)

\method{simulateData}{bernoulli}(n, size, prob, ...)

\method{simulateData}{bernoulliC}(n, size, prob, ...)
}
\arguments{
\item{...}{S3 method compatibility}

\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{size}{number of trials (zero or more).}

\item{prob}{probability of success on each trial.}
}
\description{
Random generation for some data distribution
}
