% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianMethods.R
\name{tidyBayesTest}
\alias{tidyBayesTest}
\title{bayesTest wrapper}
\usage{
tidyBayesTest(
  A_data,
  B_data,
  priors,
  n_samples = 1e+05,
  distribution = c("bernoulli", "normal", "lognormal", "poisson", "exponential",
    "uniform", "bernoulliC", "poissonC")
)
}
\arguments{
\item{A_data}{Vector of collected samples from recipe A}

\item{B_data}{Vector of collected samples from recipe B}

\item{priors}{Named vector or named list providing priors as required by the specified distribution:
\itemize{
          \item     For 'bernoulli' distribution \code{list("alpha" = val1, "beta" = val2)}
          \item     For 'normal' distribution \code{c("mu" = val1, "lambda" = val2, "alpha" = val3, "beta" = val4)}
          \item     For 'lognormal' distribution \code{c("mu" = val1, "lambda" = val2, "alpha" = val3, "beta" = val4)}
          \item     For 'poisson' distribution \code{c("shape" = val1, "rate" = val2)}
          \item     For 'exponential' distribution \code{list("shape" = val1, "rate" = val2)}
          \item     For 'uniform' distribution \code{c("xm" = val1, "alpha" = val2)}
          \item     For 'bernoulliC' distribution: same prior definitions as 'bernoulli'
          \item     For 'poissonC' distribution: same prior definitions as 'poisson'
          }

          See \link[bayesAB]{plotDistributions} or the \emph{Note} section of this help document for more info.}

\item{n_samples}{Number of posterior samples to draw. Should be large enough for the distribution to converge. 1e5 is a good rule of thumb.
Not used for closed form tests.}

\item{distribution}{Distribution of underlying A/B test data.}
}
\description{
Wrapper for \code{\link[bayesAB]{bayesTest}} which returns a tidy table instead of a messy list
}
\examples{
A_binom <- rbinom(100, 1, .5)
B_binom <- rbinom(100, 1, .6)
tidyBayesTest(A_binom, B_binom,
  priors = c("alpha" = 1, "beta" = 1),
  distribution = "bernoulli"
)
}
\seealso{
\code{\link[bayesAB]{bayesTest}}
}
