% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentistMethods.R
\name{tidyPwr2p2nTest}
\alias{tidyPwr2p2nTest}
\title{pwr.2p2n.test tidy wrapper}
\usage{
tidyPwr2p2nTest(
  h = NULL,
  n1 = NULL,
  n2 = NULL,
  sig.level = 0.05,
  power = NULL,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{h}{Effect size}

\item{n1}{Number of observations in the first sample}

\item{n2}{Number of observations in the second sample}

\item{sig.level}{Significance level (Type I error probability)}

\item{power}{Power of test (1 minus Type II error probability)}

\item{alternative}{a character string specifying the alternative hypothesis,
          must be one of "two.sided" (default), "greater" or
          "less"}
}
\description{
Wrapper for \code{\link[pwr]{pwr.2p2n.test}} which returns a tidy table instead of a messy list
}
\seealso{
\code{\link[pwr]{pwr.2p2n.test}}
}
\keyword{internal}
