% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianMethods.R
\name{runABbayesTest}
\alias{runABbayesTest}
\title{bayesTest wrapper}
\usage{
runABbayesTest(
  topScorerHitRate,
  topScorerSampleSize,
  lowScorerhitRate,
  lowScorerSampleSize,
  priors,
  distribution = c("bernoulli", "normal", "lognormal", "poisson", "exponential",
    "uniform", "bernoulliC", "poissonC"),
  nSimulatedData = 10,
  n_samples = 1e+05
)
}
\arguments{
\item{topScorerHitRate}{Numeric, hit rate for the top scorer}

\item{topScorerSampleSize}{Named numeric vector, group size for the top scorer}

\item{lowScorerhitRate}{Numeric, hit rate for the low scorer}

\item{lowScorerSampleSize}{Named numeric vector, group size for the low scorer}

\item{priors}{Named vector or named list providing priors as required by the specified distribution:
\itemize{
          \item     For 'bernoulli' distribution \code{list("alpha" = val1, "beta" = val2)}
          \item     For 'normal' distribution \code{c("mu" = val1, "lambda" = val2, "alpha" = val3, "beta" = val4)}
          \item     For 'lognormal' distribution \code{c("mu" = val1, "lambda" = val2, "alpha" = val3, "beta" = val4)}
          \item     For 'poisson' distribution \code{c("shape" = val1, "rate" = val2)}
          \item     For 'exponential' distribution \code{list("shape" = val1, "rate" = val2)}
          \item     For 'uniform' distribution \code{c("xm" = val1, "alpha" = val2)}
          \item     For 'bernoulliC' distribution: same prior definitions as 'bernoulli'
          \item     For 'poissonC' distribution: same prior definitions as 'poisson'
          }

          See \link[bayesAB]{plotDistributions} or the \emph{Note} section of this help document for more info.}

\item{nSimulatedData}{number of trials for the simulated data}

\item{n_samples}{Number of posterior samples to draw. Should be large enough for the distribution to converge. 1e5 is a good rule of thumb.
Not used for closed form tests.}
}
\value{
Returns bayesTest results in a tidy table
}
\description{
Wrapper to run multiple bayesian A/B tests on lists of A_data, B_data and priors.
}
